test_trajectory:  Read joints list from trajectory controller params
If the `hardware_interface/joints` list contains some joints
configured in a trajectory controller, the `test_trajectory` program
exits with an error:

    [ERROR]: Joints on incoming goal don't match the controller joints.
    [ INFO]: Action finished: REJECTED
    [ INFO]: TestTrajectory Finished
    [ INFO]: Shutting down.

The correct list of joints for the trajectory controller can be found
in the `<trajectory_controller_name>/joints` parameter.

Since the `<trajectory_controller_name>/follow_joint_trajectory`
parameter name is already being passed in, this patch simply shortens
that to `<trajectory_controller_name>`, and uses it to construct both
topic names.  Now the trajectory is generated for the correct number
of joints.